refactor: move scenario execution from Database to FloodAdapt class using ScenarioRunner
